xong phần auth + admin

+++ flowerShop/src/main/java/vn/flower/repositories/CategoryRepository.java
@@ -3,4 +3,5 @@ package vn.flower.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.flower.entities.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {}
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
